roachprod: implement --dry-run and --verbose
Debugging roachprod in a remote (user) environment
can be challenging. On several occassions, simply
printing all executed (cloud) commands was sufficient
to reproduce and root cause an issue.

This change implements two mutually exclusive
modes, namely `dry-run` and `verbose`. The latter
merely logs every command before it's executed.
The former suppresses execution of any command
which can modify the (cloud) infrastructure; i.e.,
`dry-run` executes read-only commands; it logs
every command, executed or not. Also note that
`dry-run` leaves temporary files like VM startup
scripts on disk, so that they can be readily examined.

Thus, `dry-run` is primarily for dumping _all_
commands without the risk of modifying anything;
`verbose` is primarily for live debugging.

Release note: None
Epic: none
Fixes: #99091

+++ pkg/roachprod/clusters_cache.go
@@ -86,10 +86,20 @@ func saveCluster(l *logger.Logger, c *cloud.Cluster) error {
 	err = errors.CombineErrors(err, tmpFile.Sync())
 	err = errors.CombineErrors(err, tmpFile.Close())
 	if err == nil {
-		err = os.Rename(tmpFile.Name(), filename)
+		if config.DryRun || config.Verbose {
+			l.Printf("exec: mv %s %s", tmpFile.Name(), filename)
+		}
+		if !config.DryRun {
+			err = os.Rename(tmpFile.Name(), filename)
+		}
 	}
 	if err != nil {
-		_ = os.Remove(tmpFile.Name())
+		if config.DryRun || config.Verbose {
+			l.Printf("exec: rm %s", tmpFile.Name())
+		}
+		if !config.DryRun {
+			_ = os.Remove(tmpFile.Name())
+		}
 		return err
 	}
 	return nil
@@ -258,5 +268,12 @@ func (localVMStorage) SaveCluster(l *logger.Logger, cluster *cloud.Cluster) erro
 // DeleteCluster is part of the local.VMStorage interface.
 func (localVMStorage) DeleteCluster(l *logger.Logger, name string) error {
 	path := clusterFilename(name)
+	if config.DryRun || config.Verbose {
+		l.Printf("exec: rm %s", path)
+	}
+	if config.DryRun {
+		return nil
+	}
+
 	return os.Remove(path)
 }
